requirements.txt
python-telegram-bot==21.5
python-telegram-bot==20.0
python-dotenv==1.0.0
